[components] Separate out runtime component and generator in dbt_project
