+++ python_modules/libraries/dagster-components/dagster_components/lib/__init__.py
@@ -4,7 +4,9 @@ _has_dagster_dbt = importlib.util.find_spec("dagster_dbt") is not None
 _has_dagster_embedded_elt = importlib.util.find_spec("dagster_embedded_elt") is not None
 
 if _has_dagster_dbt:
-    from dagster_components.lib.dbt_project import DbtProjectComponent as DbtProjectComponent
+    from dagster_components.lib.dbt_project.component import (
+        DbtProjectComponent as DbtProjectComponent,
+    )
 
 if _has_dagster_embedded_elt:
     from dagster_components.lib.sling_replication import (
